Advanced Algos added

+++ Segundo/Numerical_Computation/Practica1/succ-float.c
@@ -10,7 +10,7 @@ int main(int argc, char const *argv[])
     for (int i = 1; i < 41; i++)
     {
         x_n1 = x_n;
-        x_n = exp(-6.2 * pow(x_n,2)) - 0.5;
+        x_n =expf(-6.2 * pow(x_n,2)) - 0.5;
     }
 
     printf("\n Valor de x_40: %f\n",x_n);
